requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
supabase>=2.0.0
google-cloud-aiplatform>=1.50.0
google-generativeai>=0.3.0
groq>=0.4.0
plotly>=5.18.0
python-docx>=1.1.0
numpy>=1.24.0

# RAG 시스템 (선택 설치 - 벡터 검색 사용 시 필요)
# pip install sentence-transformers faiss-cpu
